Refactor compare_vs_tflite into library+shell
Also, drive-by change to the test names to keep them matching filenames more closely
